.767 Fix Adventurer Tab not showing (Module PluginEnabled not firing) Fix for bounties exception selecting act (collection change while enumerating) Changed InteractTag to allow ignoring IsInteractable sanity checks Fixed TakeFromStashTag QuickWithdrawl not working. Profile tag base classes now force tags to wait until trinity is ready before running (actors and scenes caches populated) Added ReloadProfileTag, HaveBountyTag. Fixes to ContainerBaseProfileBehavior Done Event firing twice & QuestStep check preventing child tags from running. Added Night's Wizard Hydra Routine Added Sequence's cursed bounty fixes Disabled auto-detection of kill-all bounty quest requirement Adjusted safe-zerg in various coroutines and made to use CombatMode Fixed Core.Grids.CanRayWalk incorrectly using RayCast check Should be generally faster to blacklist things. Added fallback for long targetting time on position based range checks (TrintiyCombat.Targeting.IsInRange(vector3)). Added remembering and reloading the last used exploration grid to preserve exploration progress. Added module to make sure data is being cleared on game restart. Adjustments to hopefully avoid some d3 crashes. Improvements to health well weighting.

+++ Framework/Avoidance/Handlers/FurnaceAvoidanceHandler.cs
@@ -25,7 +25,7 @@ namespace Trinity.Framework.Avoidance.Handlers
 
                 if (part.Type == PartType.VisualEffect)
                 {
-                    if (actor.Attributes.GetAttribute<bool>(part.Attribute, part.Power))
+                    if (actor.Attributes.GetAttribute<bool>(part.Attribute, (int)part.Power))
                     {
                         Core.Logger.Log("Power {0} on {1} ({1}) in Attribute {2}", part.Power, actor.InternalName, part.Name, part.Attribute);
                         var nodes = grid.GetRayLineAsNodes(actor.Position, MathEx.GetPointAt(actor.Position, 30f, actor.Rotation)).SelectMany(n => n.AdjacentNodes).Distinct();
